implemented routing and minor file structure changes
