+++ src/Pages/BMI.jsx
@@ -1,0 +1,5 @@
+const BMI = () => {
+  return "This is BMI page";
+};
+
+export default BMI;
